services/eda-engine/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
prophet>=1.1.0
nats-py>=0.7.0
psycopg2-binary>=2.9.0
matplotlib>=3.7.0
seaborn>=0.12.0